% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoMiner.R
\name{GOenrich3}
\alias{GOenrich3}
\title{GOenrich3}
\usage{
GOenrich3(tableSample3, tablePop3)
}
\arguments{
\item{tableSample3}{sample return value of GOtable3()}

\item{tablePop3}{population return value of GOtable3()}
}
\value{
returns a matrix with columns c("SAMPLE","POP","ENRICHMENT")
}
\description{
compute the gene enrichment in a GO category
}
\examples{
#load("data/x_tableSample3.RData")
#load("data/x_tablePop3.RData")
m<-GOenrich3(x_tableSample3,x_tablePop3)

}
