% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoMiner.R
\name{GOtable3}
\alias{GOtable3}
\title{GOtable3}
\usage{
GOtable3(hgncList, DB)
}
\arguments{
\item{hgncList}{character list of gene names}

\item{DB}{selected ontology branch of return value of subsetGOGOA}
}
\value{
returns a list whose components are c("DB","table","ngenes")
where 'DB' is the GO DB subsetted to the desired ONTOLOGY,
and 'table' is tabulation of number of occurrences of each GO
category name within the desired ONTOLOGY,
and ngenes is the total number of hgncList genes mapping to GOGOA
}
\description{
tabulate number of geneList mappings to GO categories
}
\examples{
#load("data/GOGOA3small.RData")
DB<-GOGOA3small$ontologies[["biological_process"]]

# housekeeping genes downloaded from https://housekeeping.unicamp.br/?download
#load("data/Housekeeping_Genes.RData")
hgncList<-Housekeeping_Genes[,"Gene.name"]
x<-GOtable3(hgncList,DB)

}
