% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoMiner.R
\name{hitterBeforeAfterDriver}
\alias{hitterBeforeAfterDriver}
\title{hitterBeforeAfterDriver}
\usage{
hitterBeforeAfterDriver(GOGOA3, mn = 20, mx = 200, verbose)
}
\arguments{
\item{GOGOA3}{return value of minimalistGODB::buildGODatabase()}

\item{mn}{integer minimum category size}

\item{mx}{integer maximum category size}

\item{verbose}{integer vector representing classes}
}
\value{
returns the return value of trimGOGOA3()
}
\description{
driver to invoke hitters2() and trimGOGOA3()
}
\examples{
\dontrun{
# GOGOA3.RData is too large to include in the R package
# so I need to load it from a file that is not in the package.
# Since this is in a file in my own file system, I could not
# include this as a regular example in the package.
# This example is given in full detail in the package vignette.
# You can generate GOGOA3.RData using the package 'minimalistGODB'
# or you can retrieve it from https://github.com/barryzee/GO
dir<-"/Users/barryzeeberg/personal/GODB_RDATA/goa_human/"
load(sprintf("\%s/\%s",dir,"GOGOA3_goa_human.RData"))
geneList<-GOGOA3$ontologies[["biological_process"]][1:10,"HGNC"]
GOGOA3tr<-hitterBeforeAfterDriver(GOGOA3,mn=20,mx=200,1)
}

}
