% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoMiner.R
\name{hitters2}
\alias{hitters2}
\title{hitters2}
\usage{
hitters2(GOGOA3, verbose = 1)
}
\arguments{
\item{GOGOA3}{return value of minimalistGODB::buildGODatabase()}

\item{verbose}{integer vector representing classes}
}
\value{
returns no value, but has side effect of printing information
}
\description{
determine the number of mappings for the top several genes
}
\examples{
\dontrun{
# GOGOA3.RData is too large to include in the R package
# so I need to load it from a file that is not in the package.
# Since this is in a file in my own file system, I could not
# include this as a regular example in the package.
# This example is given in full detail in the package vignette.
# You can generate GOGOA3.RData using the package 'minimalistGODB'
# or you can retrieve it from https://github.com/barryzee/GO
dir<-"/Users/barryzeeberg/personal/GODB_RDATA/goa_human/"
load(sprintf("\%s/\%s",dir,"GOGOA3_goa_human.RData"))
geneList<-GOGOA3$ontologies[["biological_process"]][1:10,"HGNC"]
hitters2(GOGOA3,1)
}

}
