% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoMiner.R
\name{human}
\alias{human}
\title{human}
\usage{
human(GOGOA3, verbose = TRUE)
}
\arguments{
\item{GOGOA3}{return value of subsetGOGOA()}

\item{verbose}{integer vector representing classes}
}
\value{
returns Boolean TRUE if species is human
}
\description{
determine if database represents human species
}
\examples{
\dontrun{
# GOGOA3.RData is too large to include in the R package
# so I need to load it from a file that is not in the package.
# Since this is in a file in my own file system, I could not
# include this as a regular example in the package.
# you can generate it using the package 'minimalistGODB'
# or you can retrieve it from https://github.com/barryzee/GO/tree/main/databases
load("/Users/barryzeeberg/personal/GODB_RDATA/goa_human/GOGOA3_goa_human.RData")
hum<-human(GOGOA3)

load("/Users/barryzeeberg/personal/GODB_RDATA/sgd/GOGOA3_sgd.RData")
hum<-human(XENOPUS,1)
}

}
