% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoMiner.R
\name{randSubsetGeneList}
\alias{randSubsetGeneList}
\title{randSubsetGeneList}
\usage{
randSubsetGeneList(geneList, ngenes)
}
\arguments{
\item{geneList}{character vector geneList}

\item{ngenes}{integer desired number of random genes}
}
\value{
returns a character vector of genes
}
\description{
retrieve n unique random genes
}
\examples{
\dontrun{
# GOGOA3.RData is too large to include in the R package
# so I need to load it from a file that is not in the package.
# Since this is in a file in my own file system, I could not
# include this as a regular example in the package.
# you can generate it using the package 'minimalistGODB'
# or you can retrieve it from https://github.com/barryzee/GO/tree/main/databases
load("/Users/barryzeeberg/personal/GODB_RDATA/goa_human/GOGOA3_goa_human.RData")
genes<-randSubsetGeneList(GOGOA3$genes[["biological_process"]],20)
}
 
}
