% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoMiner.R
\name{RCPD}
\alias{RCPD}
\title{RCPD}
\usage{
RCPD(GOGOA3, ngenes, nrand, ONT)
}
\arguments{
\item{GOGOA3}{return value of subsetGOGOA()}

\item{ngenes}{integer number of genes to randomize}

\item{nrand}{integer number of randomizations}

\item{ONT}{c("molecular_function","cellular_component","biological_process")}
}
\value{
returns a histogram of log10(p)
}
\description{
prepare a cpd of p values from randomized gene sets
}
\details{
the cpd of the randomizations is to be used for estimating
the false discovery rate (FDR) of the real sampled genes
}
\examples{
#load("data/GOGOA3small.RData")
rcpd<-RCPD(GOGOA3small,ngenes=100,nrand=10,ONT="biological_process")

}
