% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoMiner.R
\name{validHGNCSymbols}
\alias{validHGNCSymbols}
\title{validHGNCSymbols}
\usage{
validHGNCSymbols(geneList)
}
\arguments{
\item{geneList}{character vector of HGNC symbols}
}
\value{
returns list of mapping table and vector of current HGNC symbols
}
\description{
convert outdated HGNC symbols to current HGNC symbols
}
\details{
removes NA and /// from output of checkGeneSymbols()
}
\examples{
geneList<-c("FN1", "tp53", "UNKNOWNGENE","7-Sep",
 "9/7", "1-Mar", "Oct4", "4-Oct","OCT4-PG4", "C19ORF71",
  "C19orf71")
l<-validHGNCSymbols(geneList)

}
