% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoMiner.R
\name{FDR}
\alias{FDR}
\title{FDR}
\usage{
FDR(sampleList, GOGOA3, nrand, ONT)
}
\arguments{
\item{sampleList}{character vector of user-supplied genes of interest}

\item{GOGOA3}{return value of subsetGOGOA()}

\item{nrand}{integer number of randomizations}

\item{ONT}{c("molecular_function","cellular_component","biological_process")}
}
\value{
returns a list with FDR information
}
\description{
compute the false discovery rate (FDR) of the hypergeometric
p values of genes mapping to gene ontology (GO) categories
}
\examples{
#load("data/GOGOA3small.RData")
sampleList<-randSubsetGeneList(GOGOA3small$genes[["biological_process"]],10)
fdr<-FDR(sampleList,GOGOA3small,nrand=100,"biological_process")

}
