% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoMiner.R
\name{GOhypergeometric3}
\alias{GOhypergeometric3}
\title{GOhypergeometric}
\usage{
GOhypergeometric3(tableSample3, tablePop3)
}
\arguments{
\item{tableSample3}{sample return value of GOtable3()}

\item{tablePop3}{population return value of GOtable3()}
}
\value{
returns a matrix with columns c("x","m","n","k","p")
}
\description{
compute the hypergeometric p value for gene enrichment in a GO category
}
\examples{
hyper<-GOhypergeometric3(x_tableSample3,x_tablePop3)

}
