% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoMiner.R
\name{GOthresh}
\alias{GOthresh}
\title{GOthresh}
\usage{
GOthresh(m, sampleFDR, enrichThresh, countThresh, pvalThresh, fdrThresh)
}
\arguments{
\item{m}{return value of GOenrich3()}

\item{sampleFDR}{component of return value of RCPD()}

\item{enrichThresh}{numerical acceptance threshold for enrichment}

\item{countThresh}{numerical acceptance threshold for gene count}

\item{pvalThresh}{numerical acceptance threshold for pval}

\item{fdrThresh}{numerical acceptance threshold for fdr}
}
\value{
returns a subset of matrix (m joined with fdr$sampleFDR) with entries meeting all thresholds
}
\description{
retrieve lines of m that meet both enrichThresh and countThresh
}
\examples{
thresh<-GOthresh(x_m,x_fdr$sampleFDR,enrichThresh=2,countThresh=2,pvalThresh=0.1,fdrThresh=0.100)

}
