% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoMiner.R
\name{randSubsetGeneList}
\alias{randSubsetGeneList}
\title{randSubsetGeneList}
\usage{
randSubsetGeneList(geneList, ngenes)
}
\arguments{
\item{geneList}{character vector geneList}

\item{ngenes}{integer desired number of random genes}
}
\value{
returns a character vector of genes
}
\description{
retrieve n unique random genes
}
\examples{
#load("data/GOGOA3small.RData")
genes<-randSubsetGeneList(GOGOA3small$genes[["biological_process"]],20)
 
}
