% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoMiner.R
\name{runGoMinerExamples}
\alias{runGoMinerExamples}
\title{runGoMinerExamples}
\usage{
runGoMinerExamples(
  title = NULL,
  dir,
  sampleList,
  GOGOA3,
  ontology,
  enrichThresh = 2,
  countThresh = 5,
  pvalThresh = 0.1,
  fdrThresh = 0.1,
  nrand = 2,
  mn = 2,
  mx = 200,
  verbose = 1
)
}
\arguments{
\item{title}{character string descriptive title}

\item{dir}{character string full pathname to the directory acting result repository}

\item{sampleList}{character list of gene names}

\item{GOGOA3}{return value of subsetGOGOA()}

\item{ontology}{character string c("molecular_function", "cellular_component", "biological_process")}

\item{enrichThresh}{numerical acceptance threshold for enrichment}

\item{countThresh}{numerical acceptance threshold for gene count}

\item{pvalThresh}{numerical acceptance threshold for pval}

\item{fdrThresh}{numerical acceptance threshold for fdr}

\item{nrand}{numeric number of randomizations to compute FDR}

\item{mn}{integer param passed to trimGOGOA3, min size threshold for a category}

\item{mx}{integer param passed to trimGOGOA3, max size threshold for a category}

\item{verbose}{integer vector representing classes}
}
\value{
returns a list containing the return value of GoMiner()
}
\description{
driver to run GoMiner under several randomization procedures
}
\examples{
\dontrun{
# GOGOA3.RData is too large to include in the R package
# so I need to load it from a file that is not in the package.
# Since this is in a file in my own file system, I could not
# include this as a regular example in the package.
# you can generate it using the package 'minimalistGODB'
# or you can retrieve it from https://github.com/barryzee/GO/tree/main/databases
load("/Users/barryzeeberg/personal/GODB_RDATA/goa_human/GOGOA3_goa_human.RData")
ontology<-"biological_process"
t<-sort(table(GOGOA3$ontologies[[ontology]][,"HGNC"]),decreasing=TRUE)
dir<-tempdir()

sampleList<-names(t)[1:50]
title<-"hi_hitters"
hh<-runGoMinerExamples(title,dir,sampleList,GOGOA3,ontology,nrand=5)

sampleList<-names(t)[1001:1050]
title<-"hi_hitters5"
hh<-runGoMinerExamples(title,dir,sampleList,GOGOA3,ontology,nrand=5)

sampleList<-cluster52
title<-"cluster52"
hh<-runGoMinerExamples(title,dir,sampleList,GOGOA3,ontology,nrand=5)
}

}
